MVVMにリファクタリング
