+++ lib/main_page.dart
@@ -1,0 +1,90 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:mvvm_app/main_page_vm.dart';
+
+
+class MainPage extends ConsumerStatefulWidget {
+  const MainPage({Key? key, required this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  ConsumerState<MainPage> createState() => _MainPageState();
+}
+
+class _MainPageState extends ConsumerState<MainPage> {
+  final MainPageVM _vm = MainPageVM();
+
+  @override
+  void initState() {
+    super.initState();
+    _vm.setRef(ref);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
+      body: _vm.userList().when(
+        data: (user) {
+          return RefreshIndicator(
+            child: ListView.builder(
+                itemCount: user.results.length,
+                itemBuilder: (context, index) {
+                  return Column(
+                    children: [
+                      Padding(
+                        padding: const EdgeInsets.all(8.0),
+                        child: ListTile(
+                          title: Row(
+                            children: [
+                              CircleAvatar(
+                                backgroundImage: NetworkImage(user.results[index].picture.large),
+                              ),
+                              const SizedBox(width: 20.0),
+                              Expanded(
+                                child: Column(
+                                  crossAxisAlignment:
+                                  CrossAxisAlignment.start,
+                                  children: [
+                                    Text(
+                                      '${user.results[index].name.title} ${user.results[index].name.first} ${user.results[index].name.last}',
+                                      style: const TextStyle(fontSize: 17.0),
+                                      overflow: TextOverflow.ellipsis,
+                                    ),
+                                    const SizedBox(height: 10.0),
+                                    Text(
+                                      user.results[index].email.toString(),
+                                      style: const TextStyle(
+                                        fontSize: 16.0,
+                                        color: Colors.grey,
+                                        overflow: TextOverflow.ellipsis,
+                                      ),
+                                    ),
+                                  ],
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                      )
+                    ],
+                  );
+                }
+            ),
+            onRefresh: () async {
+              _vm.onRefresh();
+            },
+          );
+        },
+        error: (error, stackTrace) =>
+            Center(
+                child: Text(error.toString())
+            ),
+        loading: () => const Center(child: CircularProgressIndicator()),
+      ),
+    );
+  }
+}
